one more fix

+++ Typed/Utilities.py
@@ -5,7 +5,7 @@ this_package_path = os.path.dirname(os.path.abspath(__file__))
 sys.path.insert(0, os.path.join(this_package_path, '..'))
 
 #TODO: Cannot type variable in retic
-data = (list(map(int, [line.strip() for line in open(
+data = (list(map(float, [line.strip() for line in open(
     "utill-random-numbers.txt")])))
 rand_num = (element for element in data)
